feat: refactor API handlers for local and remote FCB file support
Introduced a new `fs_handler` module for handling local file access and updated the `http_handler` for remote file access. Enhanced the `AppState` structure to include a flag indicating whether the FCB URL is a local file. Updated the routing logic in the `create_app` function to direct requests to the appropriate handler based on the file type. This refactor improves the flexibility of the API in handling different data sources.
